+++ R/tweakPedigree.R
@@ -290,3 +290,60 @@ dropLink <- function(ped,
   }
   return(ped)
 }
+#' addPersonToTree
+#' A function to add a new person to an existing pedigree \code{data.frame}.
+#' @param ped A \code{data.frame} representing the existing pedigree.
+#' @param name Optional. A character string representing the name of the new person. If not provided, the name will be set to \code{NA}.
+#' @param sex A value representing the sex of the new person.
+#' @param momID Optional. The ID of the mother of the new person. If not provided, it will be set to \code{NA}.
+#' @param dadID Optional. The ID of the father of the new person. If not provided, it will be set to \code{NA}.
+#'@param twinID Optional. The ID of the twin of the new person. If not provided, it will be set to \code{NA}.
+#'@param personID Optional. The ID of the new person. If not provided, it will be generated as the maximum existing personID + 1.
+#'
+#' @return A \code{data.frame} with the new person added to the existing pedigree.
+#'
+addPersonToPed <- function(ped, name = NULL, sex = NULL, momID = NA, dadID = NA, twinID = NULL, personID=NULL) {
+  stopifnot(is.data.frame(ped))
+
+    # Copy structure from an existing row
+    new_row <- ped[1, , drop = FALSE]
+
+    # Blank out all values
+    new_row[1, ] <- NA
+
+    # Assign new values
+    if (!is.null(personID)) {
+      new_row$personID <- personID
+    } else {
+      # Generate a new personID based on the maximum existing personID
+        new_row$personID <- max(ped$personID, na.rm = TRUE) + 1
+    }
+    if (!is.null(name) && "name" %in% colnames(ped)) {
+      new_row$name <- name
+    } else if ("name" %in% colnames(ped)) {
+      new_row$name <- NA_character_
+    }
+    if(!is.null(twinID) && "twinID" %in% colnames(ped)) {
+      new_row$twinID <- twinID
+    } else if ("twinID" %in% colnames(ped)) {
+      new_row$twinID <- NA_integer_
+    }
+    if(!is.null(momID) && "momID" %in% colnames(ped)) {
+      new_row$momID <- momID
+    } else if ("momID" %in% colnames(ped)) {
+      new_row$momID <- NA_integer_
+    }
+    if(!is.null(dadID) && "dadID" %in% colnames(ped)) {
+      new_row$dadID <- dadID
+    } else if ("dadID" %in% colnames(ped)) {
+      new_row$dadID <- NA_integer_
+    }
+if(!is.null(sex) && "sex" %in% colnames(ped)) {
+  new_row$sex      <- sex
+  } else if ("sex" %in% colnames(ped)) {
+  new_row$sex <- NA_character_
+  }
+
+    # Append to data frame
+    rbind(ped, new_row)
+}