+++ NEWS.md
@@ -2,6 +2,8 @@
 * Added twinIDs for potter and asoiaf pedigrees
 * Added twinID to simulatePedigree function, and extended to include MZ, DZ, and SS twins.
 * Added a few more tests for simulatePedigree
+* Added function to easily add new person to a pedigree
+* Updated ASOIAF pedigree to missing parents
 
 # BGmisc 1.4.1
 * replaced print with message in all functions